+++ src/addRemove.js
@@ -1,33 +1,30 @@
-// /* eslint-disable import/no-cycle */
-// import { tasks, saveChanges, saveLocalstorage } from './drag.js';
-// import { tasksList } from './index.js';
+/* eslint-disable import/no-cycle */
+import { list, saveChanges, saveLocalstorage } from './drag.js';
+import { todoList, del } from './index.js';
 
-// const addTask = (task) => {
-//   const newTask = document.createElement('div');
-//   newTask.classList.add('mini-section');
-//   newTask.classList.add('item');
-//   newTask.setAttribute('draggable', 'true');
-//   newTask.innerHTML = `
-//   <span>
-//   <input class='check' type='checkbox' id='task-description' name='task-description' value='${task}'>${task}
-//   <i class="fas fa-ellipsis-v ellipsis" style="color: gray; float: right;"></i>
-//   </span>
-//   `;
-//   tasksList.appendChild(newTask);
-//   saveChanges();
-//   saveLocalstorage(tasks);
-//   document.location.reload(true);
-// }
+export default function addTask(task) {
+  todoList.innerHTML = '';
+  const newTask = document.createElement('div');
+  newTask.classList.add('mini-section');
+  newTask.classList.add('item');
+  newTask.setAttribute('draggable', 'true');
+  newTask.innerHTML = `
+      <span>
+      <input class='check' type='checkbox' id='task-description' name='task-description' value='${task}'>${task}
+      <i class="fas fa-ellipsis-v ellipsis" style="color: gray; float: right;"></i>
+      </span>
+      `;
+  todoList.appendChild(newTask);
+  saveChanges();
+  saveLocalstorage(list);
+  document.location.reload(true);
+}
 
-// export{addTask};
-
-// delete task function
-// function deleteTask(index){
-//   let getLocalStorageData = localStorage.getItem("list");
-//   list = JSON.parse(getLocalStorageData);
-//   list.splice(index, 1); //delete or remove the li
-//   localStorage.setItem("New Todo", JSON.stringify(list));
-//   displayList(list); //call the displayList function
-// }
-
-// export {deleteTask};
+export const deleteItem = () => {
+  del.addEventListener('click', (e) => {
+    e.preventDefault();
+    e.target.parentElement.parentElement.remove();
+    saveChanges();
+    document.location.reload(true);
+  });
+};
