+++ src/drag.js
@@ -2,8 +2,9 @@
 /* eslint-disable import/no-mutable-exports,  */
 /* eslint-disable import/no-cycle */
 /* eslint-disable no-loop-func */
+import { clearAll } from './index.js';
+import addTask from './addRemove.js';
 
-import { todoList, items, del, clearAll } from './index.js';
 export let list = [];
 
 if (localStorage.getItem('list')) {
@@ -53,24 +54,6 @@ export const saveChanges = () => {
   }
 };
 
-const addTask = (task) => {
-    const newTask = document.createElement('div');
-    newTask.classList.add('mini-section');
-    newTask.classList.add('item');
-    newTask.setAttribute('draggable', 'true');
-    newTask.innerHTML = `
-    <span>
-    <input class='check' type='checkbox' id='task-description' name='task-description' value='${task}'>${task}
-    <i class="fas fa-ellipsis-v ellipsis" style="color: gray; float: right;"></i>
-    </span>
-    `;
-    todoList.appendChild(newTask);
-    saveChanges();
-    saveLocalstorage(list);
-    document.location.reload(true);
-};
-
-
 export const inputListener = (input) => {
   input.addEventListener('keypress', (e) => {
     if (e.key === 'Enter') {
@@ -116,59 +99,6 @@ export const dragOver = (container) => {
       container.insertBefore(draggable, afterElement);
     }
   });
-};
-
-export const editListener = () => {
-  for (let i = 0; i < items.length; i += 1) {
-    items[i].addEventListener('dblclick', () => {
-      const oldValue = items[i].children[0].children[0];
-      const oldElement = items[i].children[0].children[0].nextSibling.nextSibling;
-      oldElement.remove();
-      items[i].children[0].children[0].nextSibling.remove();
-      const newInput = document.createElement('input');
-      const icon = document.createElement('i');
-      icon.classList.add('fa', 'fa-trash', 'trash');
-      items[i].firstChild.nextSibling.removeChild(items[i].children[0].children[0]);
-      newInput.placeholder = oldValue.value;
-      items[i].children[0].insertBefore(newInput, items[i].children[0].children[0]);
-      items[i].children[0].insertBefore(icon, items[i].children[0].children[0]);
-
-      newInput.addEventListener('keypress', (e) => {
-        if (e.key === 'Enter') {
-          if (newInput.value.trim() === '') {
-            newInput.classList.add('invalid');
-          } else {
-            items[i].children[0].id = i;
-            newInput.setAttribute('description', newInput.value);
-            items[i].firstChild.nextSibling.removeChild(items[i].children[0].children[0]);
-
-            items[i].firstChild.nextSibling.innerHTML = `
-           <input class='check' type='checkbox' id='task-description' name='task-description' value='${newInput.value}'>${newInput.value}
-           <i class="fas fa-ellipsis-v ellipsis" style="color: gray; float: right;"></i>
-          `;
-            list[i].description = newInput.value;
-            list[i].completed = false;
-            saveChanges();
-          }
-        }
-      });
-
-      icon.addEventListener('click', (e) => {
-        e.target.parentElement.parentElement.remove();
-        saveChanges();
-        document.location.reload(true);
-      });
-    });
-  }
-};
-
-export const deleteItem =  () => {
-  del.addEventListener('click', (e) => {
-    e.preventDefault();
-    del.parentNode.parentNode.removeChild(del.parentNode);
-    saveLocalstorage();
-    document.location.reload(true);
-});
 };
 
 export const deleteAll = () => {
